refactored and bug fixed
