+++ GUI/src/main/kotlin/ui/components/FileSelectorButton.kt
@@ -65,3 +65,18 @@ fun openFileChooserAndGetPath(): String? {
     val result = fileChooser.showOpenDialog(null)
     return if (result == JFileChooser.APPROVE_OPTION) fileChooser.selectedFile.absolutePath else null
 }
+
+/**
+ * Opens a file chooser dialog and returns the selected file path.
+ *
+ * @return The selected file path, or null if no file was selected.
+ */
+fun openSaveChooserAndGetPath(): String? {
+    val fileChooser = JFileChooser()
+    val result = fileChooser.showSaveDialog(null)
+    return if (result == JFileChooser.APPROVE_OPTION) {
+        fileChooser.selectedFile.absolutePath
+    } else {
+        null
+    }
+}
