+++ GUI/src/main/kotlin/ui/components/ProjectManager.kt
@@ -17,7 +17,6 @@ import com.fasterxml.jackson.databind.module.SimpleModule
 import com.fasterxml.jackson.module.kotlin.readValue
 import models.AppState
 import java.io.File
-import javax.swing.JFileChooser
 
 // singleton
 object AppConfig {
@@ -73,16 +72,6 @@ fun projectMenu(
                 Text("Save Project", fontSize = MaterialTheme.typography.body2.fontSize)
             }
         }
-    }
-}
-
-fun openSaveChooserAndGetPath(): String? {
-    val fileChooser = JFileChooser()
-    val result = fileChooser.showSaveDialog(null)
-    return if (result == JFileChooser.APPROVE_OPTION) {
-        fileChooser.selectedFile.absolutePath
-    } else {
-        null
     }
 }
